refactor: deepen the Queue live-sync IPC boundary (follow-up to #351)
Four playlists.svelte.ts store methods paired a DB-row mutation with a
separate call to keep the audio player's live in-memory queue in sync.
Merges each pair into one backend command, following the add_songs_to_queue
precedent from #346:

- remove_from_playlist and reorder_playlist_item_by_uuid now always cascade
  into the live queue — both are UUID-keyed, and UUIDs never collide across
  playlists, so the cascade is a safe no-op whenever the live queue doesn't
  hold that row
- reorder_playlist_item and clear_playlist cascade only when the target
  playlist is the Queue (checked server-side via PlaylistManager::queue());
  unlike the UUID-keyed calls, an index pair or a full clear has no
  self-limiting scope, so applying either unconditionally would corrupt
  live playback when editing an unrelated playlist
- deleted replaceQueueTracks, which turned out to have zero callers

remove_songs_from_player_playlist, reorder_player_item_by_uuid,
reorder_player_playlist_items, and clear_player_playlist are now
unreachable from the frontend and are removed along with their lib.rs
registrations.
