+++ src-tauri/src/commands/player.rs
@@ -492,63 +492,3 @@ pub async fn add_songs_to_queue(
     Ok(())
 }
 
-/// Remove items from the live in-memory player playlist by uuid, so songs
-/// removed from the Queue's DB playlist stop being up-next during playback.
-/// Called by the frontend after `remove_from_playlist` has deleted the rows
-/// from the DB (see #262).
-#[tauri::command]
-pub async fn remove_songs_from_player_playlist(
-    uuids: Vec<String>,
-    app: tauri::AppHandle,
-    state: State<'_, AppState>,
-) -> Result<(), String> {
-    use tauri::Emitter;
-    let mut player = state.player.lock().await;
-    player.remove_songs_from_playlist_items(&uuids);
-    let playback_state = player.get_state().await;
-    let _ = app.emit("playback-state", playback_state);
-    Ok(())
-}
-
-#[tauri::command]
-pub async fn reorder_player_playlist_items(
-    from_index: usize,
-    to_index: usize,
-    app: tauri::AppHandle,
-    state: State<'_, AppState>,
-) -> Result<(), String> {
-    use tauri::Emitter;
-    let mut player = state.player.lock().await;
-    player.reorder_playlist_items(from_index, to_index);
-    let playback_state = player.get_state().await;
-    let _ = app.emit("playback-state", playback_state);
-    Ok(())
-}
-
-#[tauri::command]
-pub async fn reorder_player_item_by_uuid(
-    source_uuid: String,
-    target_uuid: String,
-    app: tauri::AppHandle,
-    state: State<'_, AppState>,
-) -> Result<(), String> {
-    use tauri::Emitter;
-    let mut player = state.player.lock().await;
-    player.reorder_playlist_item_by_uuid(&source_uuid, &target_uuid);
-    let playback_state = player.get_state().await;
-    let _ = app.emit("playback-state", playback_state);
-    Ok(())
-}
-
-#[tauri::command]
-pub async fn clear_player_playlist(
-    app: tauri::AppHandle,
-    state: State<'_, AppState>,
-) -> Result<(), String> {
-    use tauri::Emitter;
-    let mut player = state.player.lock().await;
-    player.clear_playlist_items();
-    let playback_state = player.get_state().await;
-    let _ = app.emit("playback-state", playback_state);
-    Ok(())
-}
