+++ src-tauri/src/lib.rs
@@ -673,10 +673,6 @@ pub fn run() {
             commands::player::open_and_play,
             commands::player::get_startup_file,
             commands::player::add_songs_to_queue,
-            commands::player::remove_songs_from_player_playlist,
-            commands::player::reorder_player_playlist_items,
-            commands::player::reorder_player_item_by_uuid,
-            commands::player::clear_player_playlist,
             commands::player::pause,
             commands::player::resume,
             commands::player::stop,
